Don't sign the HMAC in key exchange
Note: Prior to this commit, the HMAC was being signed with the origin
private key and verified by the peer using the origin public key. This
was an experimental step and is removed as of this commit. For more
info on why this change was made, refer the changelogs or email the
maintainers.

* crypto/aead_ossl.c: Allow the caller to query output buffer size.
* crypto/kex_*: Remove signature on HMAC (this field is also removed
  from the kex_peer_share_st struct now that it is no longer needed).